Rename lsp_client module to jsonrpc_forwarder
Still not sure this is the best name. :P But maybe better.
